+++ patch-chromium-swt/src/main/java/org/eclipse/swt/browser/WindowEvent.java
@@ -1,0 +1,24 @@
+/*******************************************************************************
+ * Copyright (c) 2022 EquoTech, Inc. and others.
+ *
+ * This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License 2.0
+ * which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     EquoTech, Inc. - initial API and implementation
+ *******************************************************************************/
+package org.eclipse.swt.browser;
+
+import org.eclipse.swt.widgets.Widget;
+
+public class WindowEvent extends com.equo.chromium.swt.WindowEvent {
+
+	public WindowEvent(Widget widget) {
+		super(widget);
+		browser = (Browser) widget;
+	}
+}